Math1f/target/clover/src-instrumented/org/apache/commons/math3/optimization/fitting/PolynomialFitter.java
/* $$ This file has been instrumented by Clover 4.4.1#2019101123313948 $$ *//*
 * Licensed to the Apache Software Foundation (ASF) under one or more
 * contributor license agreements.  See the NOTICE file distributed with
 * this work for additional information regarding copyright ownership.
 * The ASF licenses this file to You under the Apache License, Version 2.0
 * (the "License"); you may not use this file except in compliance with
 * the License.  You may obtain a copy of the License at
 *
 *      http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */

package org.apache.commons.math3.optimization.fitting;

import org.apache.commons.math3.analysis.polynomials.PolynomialFunction;
import org.apache.commons.math3.optimization.DifferentiableMultivariateVectorOptimizer;

/**
 * Polynomial fitting is a very simple case of {@link CurveFitter curve fitting}.
 * The estimated coefficients are the polynomial coefficients (see the
 * {@link #fit(double[]) fit} method).
 *
 * @version $Id$
 * @deprecated As of 3.1 (to be removed in 4.0).
 * @since 2.0
 */
@Deprecated
public class PolynomialFitter extends CurveFitter<PolynomialFunction.Parametric> {public static class __CLR4_4_1yrhyrhlb90pa3m{public static com_atlassian_clover.CoverageRecorder R;public static com_atlassian_clover.CloverProfile[] profiles = { };@java.lang.SuppressWarnings("unchecked") public static <I, T extends I> I lambdaInc(final int i,final T l,final int si){java.lang.reflect.InvocationHandler h=new java.lang.reflect.InvocationHandler(){public java.lang.Object invoke(java.lang.Object p,java.lang.reflect.Method m,java.lang.Object[] a) throws Throwable{R.inc(i);R.inc(si);try{return m.invoke(l,a);}catch(java.lang.reflect.InvocationTargetException e){throw e.getCause()!=null?e.getCause():new RuntimeException("Clover failed to invoke instrumented lambda",e);}}};return (I)java.lang.reflect.Proxy.newProxyInstance(l.getClass().getClassLoader(),l.getClass().getInterfaces(),h);}static{com_atlassian_clover.CoverageRecorder _R=null;try{com_atlassian_clover.CloverVersionInfo.An_old_version_of_clover_is_on_your_compilation_classpath___Please_remove___Required_version_is___4_4_1();if(2019101123313948L!=com_atlassian_clover.CloverVersionInfo.getBuildStamp()){com_atlassian_clover.Clover.l("[CLOVER] WARNING: The Clover version used in instrumentation does not match the runtime version. You need to run instrumented classes against the same version of Clover that you instrumented with.");com_atlassian_clover.Clover.l("[CLOVER] WARNING: Instr=4.4.1#2019101123313948,Runtime="+com_atlassian_clover.CloverVersionInfo.getReleaseNum()+"#"+com_atlassian_clover.CloverVersionInfo.getBuildStamp());}R=com_atlassian_clover.Clover.getNullRecorder();_R=com_atlassian_clover.Clover.getNullRecorder();_R=com_atlassian_clover.Clover.getRecorder("\u002f\u0068\u006f\u006d\u0065\u002f\u0073\u0065\u0070\u0065\u0068\u0072\u002f\u006d\u0061\u0073\u0074\u0065\u0072\u0073\u002f\u0074\u0065\u0073\u0074\u005f\u0067\u0065\u006e\u0065\u0072\u0061\u0074\u0069\u006f\u006e\u002f\u0061\u006c\u006c\u005f\u0070\u0072\u006f\u006a\u0065\u0063\u0074\u0073\u005f\u0063\u006f\u0076\u0065\u0072\u0061\u0067\u0065\u002f\u004d\u0061\u0074\u0068\u0031\u0066\u002f\u0074\u0061\u0072\u0067\u0065\u0074\u002f\u0063\u006c\u006f\u0076\u0065\u0072\u002f\u0063\u006c\u006f\u0076\u0065\u0072\u002e\u0064\u0062",1670137421344L,8589935092L,45065,profiles,new java.lang.String[]{"clover.distributed.coverage",null});}catch(java.lang.SecurityException e){java.lang.System.err.println("[CLOVER] FATAL ERROR: Clover could not be initialised because it has insufficient security privileges. Please consult the Clover documentation on the security policy file changes required. ("+e.getClass()+":"+e.getMessage()+")");}catch(java.lang.NoClassDefFoundError e){java.lang.System.err.println("[CLOVER] FATAL ERROR: Clover could not be initialised. Are you sure you have Clover in the runtime classpath? ("+e.getClass()+":"+e.getMessage()+")");}catch(java.lang.Throwable t){java.lang.System.err.println("[CLOVER] FATAL ERROR: Clover could not be initialised because of an unexpected error. ("+t.getClass()+":"+t.getMessage()+")");}R=_R;}}public static final com_atlassian_clover.TestNameSniffer __CLR4_4_1_TEST_NAME_SNIFFER=com_atlassian_clover.TestNameSniffer.NULL_INSTANCE;
    /** Polynomial degree.
     * @deprecated
     */
    @Deprecated
    private final int degree;

    /**
     * Simple constructor.
     * <p>The polynomial fitter built this way are complete polynomials,
     * ie. a n-degree polynomial has n+1 coefficients.</p>
     *
     * @param degree Maximal degree of the polynomial.
     * @param optimizer Optimizer to use for the fitting.
     * @deprecated Since 3.1 (to be removed in 4.0). Please use
     * {@link #PolynomialFitter(DifferentiableMultivariateVectorOptimizer)} instead.
     */
    @Deprecated
    public PolynomialFitter(int degree, final DifferentiableMultivariateVectorOptimizer optimizer) {
        super(optimizer);__CLR4_4_1yrhyrhlb90pa3m.R.inc(45054);try{__CLR4_4_1yrhyrhlb90pa3m.R.inc(45053);
        __CLR4_4_1yrhyrhlb90pa3m.R.inc(45055);this.degree = degree;
    }finally{__CLR4_4_1yrhyrhlb90pa3m.R.flushNeeded();}}

    /**
     * Simple constructor.
     *
     * @param optimizer Optimizer to use for the fitting.
     * @since 3.1
     */
    public PolynomialFitter(DifferentiableMultivariateVectorOptimizer optimizer) {
        super(optimizer);__CLR4_4_1yrhyrhlb90pa3m.R.inc(45057);try{__CLR4_4_1yrhyrhlb90pa3m.R.inc(45056);
        __CLR4_4_1yrhyrhlb90pa3m.R.inc(45058);degree = -1; // To avoid compilation error until the instance variable is removed.
    }finally{__CLR4_4_1yrhyrhlb90pa3m.R.flushNeeded();}}

    /**
     * Get the polynomial fitting the weighted (x, y) points.
     *
     * @return the coefficients of the polynomial that best fits the observed points.
     * @throws org.apache.commons.math3.exception.ConvergenceException
     * if the algorithm failed to converge.
     * @deprecated Since 3.1 (to be removed in 4.0). Please use {@link #fit(double[])} instead.
     */
    @Deprecated
    public double[] fit() {try{__CLR4_4_1yrhyrhlb90pa3m.R.inc(45059);
        __CLR4_4_1yrhyrhlb90pa3m.R.inc(45060);return fit(new PolynomialFunction.Parametric(), new double[degree + 1]);
    }finally{__CLR4_4_1yrhyrhlb90pa3m.R.flushNeeded();}}

    /**
     * Get the coefficients of the polynomial fitting the weighted data points.
     * The degree of the fitting polynomial is {@code guess.length - 1}.
     *
     * @param guess First guess for the coefficients. They must be sorted in
     * increasing order of the polynomial's degree.
     * @param maxEval Maximum number of evaluations of the polynomial.
     * @return the coefficients of the polynomial that best fits the observed points.
     * @throws org.apache.commons.math3.exception.TooManyEvaluationsException if
     * the number of evaluations exceeds {@code maxEval}.
     * @throws org.apache.commons.math3.exception.ConvergenceException
     * if the algorithm failed to converge.
     * @since 3.1
     */
    public double[] fit(int maxEval, double[] guess) {try{__CLR4_4_1yrhyrhlb90pa3m.R.inc(45061);
        __CLR4_4_1yrhyrhlb90pa3m.R.inc(45062);return fit(maxEval, new PolynomialFunction.Parametric(), guess);
    }finally{__CLR4_4_1yrhyrhlb90pa3m.R.flushNeeded();}}

    /**
     * Get the coefficients of the polynomial fitting the weighted data points.
     * The degree of the fitting polynomial is {@code guess.length - 1}.
     *
     * @param guess First guess for the coefficients. They must be sorted in
     * increasing order of the polynomial's degree.
     * @return the coefficients of the polynomial that best fits the observed points.
     * @throws org.apache.commons.math3.exception.ConvergenceException
     * if the algorithm failed to converge.
     * @since 3.1
     */
    public double[] fit(double[] guess) {try{__CLR4_4_1yrhyrhlb90pa3m.R.inc(45063);
        __CLR4_4_1yrhyrhlb90pa3m.R.inc(45064);return fit(new PolynomialFunction.Parametric(), guess);
    }finally{__CLR4_4_1yrhyrhlb90pa3m.R.flushNeeded();}}
}
